fix: alert icon 수정
