--- src/utils/infoAlert.js
@@ -1,10 +1,0 @@
-import Swal from 'sweetalert2';
-
-export const infoAlert = ({ title }) => {
-  Swal.fire({
-    icon: 'success',
-    title: title,
-    showConfirmButton: false,
-    timer: 1500,
-  });
-};
